refactor: simplify reader slice.
