+++ lib/features/reader/models/reader_info.dart
