+++ lib/features/reader/models/toolbar_info.dart
