--- lib/features/reader/presentation/presentation.dart
@@ -1,21 +1,0 @@
-// Models.
-export 'models/reader_info.dart';
-export 'models/reader_page_info.dart';
-export 'models/toolbar_info.dart';
-
-// Pages.
-export 'pages/reader_page.dart';
-
-// Provides.
-export 'providers/reader_loading_family.dart';
-export 'providers/reader_page_family.dart';
-export 'providers/reader_family.dart';
-export 'providers/toolbar_provider.dart';
-
-// Widgets.
-export 'widgets/chapter_page.dart';
-export 'widgets/reader_body.dart';
-export 'widgets/reader_bottom_bar.dart';
-export 'widgets/reader_theme.dart';
-export 'widgets/reader_view.dart';
-export 'widgets/settings_sheet.dart';
