+++ lib/features/reader/views/reader_page.dart
@@ -2,7 +2,9 @@ import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:nacht/common/common.dart';
 
-import '../presentation.dart';
+import '../models/reader_info.dart';
+import '../providers/reader_loading_family.dart';
+import '../widgets/reader_view.dart';
 
 class ReaderPage extends ConsumerWidget {
   const ReaderPage({
